% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotImpliedVol.R
\name{plotImpliedVol}
\alias{plotImpliedVol}
\title{Calculate and plot implied volatility using Black Scholes model}
\usage{
plotImpliedVol(strike = c(50, 20), type = c("C", "P"),
  optionPrice = c(1.62, 0.01), futurePrice = c(48.03, 48.03),
  time_to_expiry = c(0.1423, 0.1423))
}
\arguments{
\item{strike}{price}

\item{Option}{price}

\item{future}{price}
}
\value{
calcualted sigma and plots
}
\description{
Calculate and plot implied volatility using Black Scholes model
}

